refactor: reorganize import paths and update task use cases for consistency
