+++ lib/features/home/domain/entities/nav_item.dart
