+++ lib/features/home/domain/repositories/i_navigation_repository.dart
@@ -1,0 +1,10 @@
+import 'package:uptodo/features/home/domain/entities/nav_item.dart';
+
+/// Repository for navigation items
+abstract class INavigationRepository {
+  /// Get all navigation items
+  List<NavItem> getNavigationItems();
+
+  /// Handle logout
+  Future<bool> handleLogout();
+}
