--- lib/features/home/index_screen/domain/usecases/handle_logout.dart
@@ -1,18 +1,0 @@
-import 'package:injectable/injectable.dart';
-import 'package:uptodo/core/utils/use_case.dart';
-import 'package:uptodo/features/home/index_screen/domain/repositories/i_navigation_repository.dart';
-
-/// HandleLogout.dart
-@injectable
-class HandleLogout extends UseCase<bool, void> {
-  /// constructor
-  HandleLogout(this.repository);
-
-  /// repository
-  final INavigationRepository repository;
-
-  @override
-  Future<bool> run(_) async {
-    return repository.handleLogout();
-  }
-}
