+++ lib/features/task_details/data/repositories/task_details_repository_impl.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:injectable/injectable.dart';
+import 'package:uptodo/features/home/domain/repositories/task_details_repository.dart';
 import 'package:uptodo/features/task_details/data/database/tasks_database_module.dart';
 import 'package:uptodo/features/task_details/data/model/task_analytics_data/task_analytics_data.dart';
+import 'package:uptodo/features/task_details/data/model/task_event_types.dart';
 import 'package:uptodo/features/task_details/data/model/task_model/task_model.dart';
-import 'package:uptodo/features/task_details/domain/repositories/task_details_repository.dart';
-import 'package:uptodo/shared/model/shred_enums.dart';
 
 /// TaskDetailsRepositoryImpl
 @Injectable(as: TaskDetailsRepository)
@@ -47,28 +47,8 @@ class TaskDetailsRepositoryImpl implements TaskDetailsRepository {
   }
 
   @override
-  Stream<TaskModel> watchTask() {
-    return _databaseModule.watchTaskChanges().map(
-          (task) => TaskModel(
-            taskId: task?.taskId.toString(),
-            title: task?.title ?? '',
-            description: task?.description ?? '',
-            subTaskId: task?.subTaskId,
-            priorityId: task?.priorityId ?? 0,
-            taskTime: task != null
-                ? TimeOfDay.fromDateTime(task.taskTime)
-                : const TimeOfDay(hour: 0, minute: 0),
-            categoryId: task?.categoryId ?? 0,
-            status: task != null ? task.status : TaskStatus.pending,
-            taskDate: task != null
-                ? DateTime(
-                    task.taskTime.year,
-                    task.taskTime.month,
-                    task.taskTime.day,
-                  )
-                : DateTime.now(),
-          ),
-        );
+  Stream<TaskTableEvents> watchTask() {
+    return _databaseModule.watchTaskChanges();
   }
 
   @override
@@ -79,5 +59,15 @@ class TaskDetailsRepositoryImpl implements TaskDetailsRepository {
   }) {
     // TODO: implement watchTask
     throw UnimplementedError();
+  }
+
+  /// this method will return if there is any task available
+  /// for today,yesterday or tomorrow.
+  /// will be false, if there is none
+  /// will be true if there is any task available
+  @override
+  Future<bool> loadTaskSummary(DateTime date) async {
+    final taskSummary = await _databaseModule.hasTasksInThreeDayRange(date);
+    return taskSummary.values.any((hasTask) => hasTask);
   }
 }
