+++ lib/features/task_details/domain/repositories/task_create_update_repository.dart
@@ -15,7 +15,7 @@ abstract class TaskCreateUpdateRepository {
   );
 
   /// delete task
-  Future<bool> deleteTask(
+  Future<void> deleteTask(
     int id,
   );
 }
